Resolved import errors, file renames,and deleted unnecesary files
