+++ src/autoplex/auto/rss/flows.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass, field
 
-from atomate2.castep.jobs.base import BaseCastepMaker
+from autoplex.castep.jobs import BaseCastepMaker
 from atomate2.forcefields.jobs import ForceFieldStaticMaker
 from atomate2.vasp.jobs.base import BaseVaspMaker
 from atomate2.vasp.jobs.core import StaticMaker
