--- src/autoplex/castep/job/__init__.py
@@ -1,1 +1,0 @@
-"Jobs for CASTEP Workflow"
