--- src/autoplex/castep/jobs/__init__.py
@@ -1,1 +1,0 @@
-"Jobs for CASTEP Workflow"
