+++ src/autoplex/data/common/utils.py
@@ -28,10 +28,8 @@ from scipy.sparse.linalg import LinearOperator, svds
 from sklearn.model_selection import StratifiedShuffleSplit
 from threadpoolctl import threadpool_limits
 from atomate2.forcefields.jobs import ForceFieldStaticMaker
-from autoplex.settings import RssConfig
-from autoplex.auto.rss.flows import RssMaker
-from jobflow import Flow, run_locally
-from autoplex.castep.jobs.base import BaseCastepMaker
+from jobflow import Flow
+from autoplex.castep.jobs import BaseCastepMaker
 from ase.calculators.castep import create_castep_keywords
 import logging
 import os
